Factored decapodes part into src and test. Again had to change AMR name to amr. Changed my_parse_decapode function name as well.
