+++ src/SyntacticModels.jl
@@ -2,7 +2,7 @@ module SyntacticModels
 
 include("SyntacticModelsBase.jl")
 include("amr_it.jl")
-# include("decapodes_it.jl")
+include("decapodes_it.jl")
 include("uwd_it.jl")
 # include("composite_models_it.jl")
