+++ test/Project.toml
@@ -11,3 +11,4 @@ Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 StructTypes = "856f2bd8-1eba-4b0a-8007-ebc267875bd4"
 SyntacticModels = "22bb929c-8bcf-4852-b455-eb3e1675e09c"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+Unicode = "4ec0a83e-493e-50e2-b9ac-8f72acf5a8f5"
